+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Piano DMP & Content SDK for Android
 
 ## v2.5.1-SNAPSHOT
+* Add storing for `fallback user id`
 * Updated dependencies:
     - org.jetbrains.kotlinx:kotlinx-coroutines-core [1.7.3 -> 1.8.0]
       https://github.com/Kotlin/kotlinx.coroutines